+++ src/test/java/it/gov/pagopa/register/service/operation/ProductFileServiceTest.java
@@ -1,17 +1,49 @@
 package it.gov.pagopa.register.service.operation;
 
+import it.gov.pagopa.register.connector.storage.FileStorageClient;
+import it.gov.pagopa.register.constants.AssetRegisterConstants;
+import it.gov.pagopa.register.constants.enums.UploadCsvStatus;
+import it.gov.pagopa.register.dto.operation.FileReportDTO;
+import it.gov.pagopa.register.dto.operation.ProductFileResponseDTO;
+import it.gov.pagopa.register.dto.operation.ProductFileResult;
+import it.gov.pagopa.register.dto.operation.ValidationResultDTO;
+import it.gov.pagopa.register.exception.operation.ReportNotFoundException;
+import it.gov.pagopa.register.model.operation.ProductFile;
+import it.gov.pagopa.register.repository.operation.ProductFileRepository;
+import it.gov.pagopa.register.utils.CsvUtils;
+import org.apache.commons.csv.CSVRecord;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.MockedStatic;
+import org.mockito.MockitoAnnotations;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.mock.web.MockMultipartFile;
+import org.springframework.web.multipart.MultipartFile;
+
+import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.util.*;
+
+import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
 class ProductFileServiceTest {
 
- /* @Mock
+  @Mock
   ProductFileRepository productFileRepository;
   @Mock
   FileStorageClient fileStorageClient;
   @Mock
-  ProductFileValidator productFileValidator;
+  ProductFileValidatorService productFileValidator;
 
   private ProductFileService productFileService;
 
@@ -43,7 +75,6 @@ class ProductFileServiceTest {
     assertEquals(2, resp.getPageSize());
     assertEquals(2, resp.getTotalElements());
     assertEquals(1, resp.getTotalPages());
-
     verify(productFileRepository).findByOrganizationIdAndUploadStatusNot(org, UploadCsvStatus.FORMAL_ERROR.name(), page);
   }
 
@@ -178,7 +209,7 @@ class ProductFileServiceTest {
     when(productFileValidator.validateFile(any(),anyString(),anyList(),anyInt())).thenReturn(validationResultDTO);
     ProductFileResult res = productFileService.processFile(file, "cat","org","user");
     assertEquals("KO", res.getStatus());
-    assertEquals("TEST", res.getErrorKey());
+    assertEquals("HEADE", res.getErrorKey());
   }
 
 
@@ -241,7 +272,7 @@ class ProductFileServiceTest {
       ProductFileResult result = productFileService.processFile(file, "cookinghobs", "org1", "user1");
 
       assertEquals("KO", result.getStatus());
-      assertEquals(AssetRegisterConstant.UploadKeyConstant.REPORT_FORMAL_FILE_ERROR_KEY, result.getErrorKey());
+      assertEquals(AssetRegisterConstants.UploadKeyConstant.REPORT_FORMAL_FILE_ERROR_KEY, result.getErrorKey());
       assertEquals("123", result.getProductFileId());
     }
   }
@@ -278,7 +309,7 @@ class ProductFileServiceTest {
 
   @Test
   void whenAllValid_thenReturnOk()  {
-    MultipartFile file = createMockFile();
+    MultipartFile file = mock(MultipartFile.class);
     CSVRecord rec = mock(CSVRecord.class);
 
     try (MockedStatic<CsvUtils> mocked = mockStatic(CsvUtils.class)) {
@@ -308,9 +339,8 @@ class ProductFileServiceTest {
       assertEquals("OK", res.getStatus());
       assertNull(res.getErrorKey());
     } catch (IOException e) {
-        throw new RuntimeException(e);
+      throw new RuntimeException(e);
     }
   }
 
-  */
 }
