+++ src/test/java/it/gov/pagopa/register/service/operation/ProductServiceTest.java
@@ -25,7 +25,7 @@ class ProductServiceTest {
   private ProductRepository productRepository;
 
   @InjectMocks
-  private ProductServiceImpl productService;
+  private ProductService productService;
 
   @BeforeEach
   void setUp() {
