+++ src/test/java/it/gov/pagopa/register/service/role/RolePermissionServiceTest.java
@@ -15,7 +15,7 @@ import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -34,7 +34,7 @@ class RolePermissionServiceTest {
     @Autowired
     RolePermissionService rolePermissionService;
 
-    @MockBean
+    @MockitoBean
     RolePermissionRepository rolePermissionRepository;
 
     @Test
